moving drawing function in sep script and reuse them
